+++ layout/formSessionMonitor.php
@@ -41,9 +41,10 @@
   </script>
 
    
-<body>
+
     
 </body>
+<?php require_once 'footer.php';?>
 </html>