+++ layout/formVersion.php
@@ -1,0 +1,57 @@
+
+
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="icon" type="image/x-icon" href="public/logo.png">
+  <title>Versiones</title>
+  <link rel="stylesheet" href="style/styleSession.css">
+
+
+</head>
+<body>
+    
+
+ 
+ 
+    <form method="post" action="controller/controllerChangePassSession.php">
+
+   
+
+    <div class="form-group">
+    <label for="nkeyword">## [1.0.5-Beta] - 2023-07-26
+- Visualizador de versiones.</label>
+    </div> 
+    <div class="form-group">
+    <label for="nkeyword">## [1.0.4-Beta] - 2023-07-26
+- Homologación de footer.</label>
+    </div> 
+    <div class="form-group">
+    <label for="nkeyword">## [1.0.3-Beta] - 2023-07-25
+- Homologación de interfaz, textos,botones.</label>
+    </div> 
+    <div class="form-group">
+    <label for="nkeyword">## [1.0.2-Beta] - 2023-07-25
+- Homologación de modales, textos.</label>
+    </div> 
+    <div class="form-group">
+    <label for="nkeyword">## [1.0.1-Beta] - 2023-07-25
+- Homologación de layouts, textos.</label>
+    </div> 
+
+    <div class="form-group">
+    <label for="nkeyword">## [1.0.0-Beta] - 2023-07-24
+- SISTEMA BASE</label>
+    </div> 
+
+
+       </form>
+
+ 
+
+ 
+</body>
+</html>