+++ view/viewSessionSuperAdmin.php
@@ -1,6 +1,8 @@
 <?php
 require_once 'layout/headerSuperAdmin.php';
+
 require_once 'layout/formSession.php';
+
 require_once 'modal/modalProfile.php';
 
 require_once 'modal/modalUsersSuperAdmin.php';
@@ -41,6 +43,7 @@ require_once 'modal/modalRemind.php';
 require_once 'modal/modalPostRemind.php';
 
 require_once 'modal/modalChangePassSession.php';
+
 
 
 ?>