+++ view/viewVersion.php
@@ -1,0 +1,5 @@
+<?php
+require_once 'controller/recharge.php';
+require_once 'layout/formVersion.php';
+
+?>