+++ changelog.md
@@ -1,10 +1,16 @@
 # Changelog
 
+## [1.0.5-Beta] - 2023-07-26
+- Visualizador de versiones.
+
+## [1.0.4-Beta] - 2023-07-26
+- Homologación de footer.
+
 ## [1.0.3-Beta] - 2023-07-25
 - Homologación de interfaz, textos,botones.
 
 ## [1.0.2-Beta] - 2023-07-25
-- Homologación de modales, textos..
+- Homologación de modales, textos.
 
 <<<<<<< HEAD
 =======